chore: fix validation refs
